feat: only have a `defer()` method on events that can be deferred (#1122)
There are two goals for this PR:

* Prevent charms from calling `defer()` on events where that makes no sense
* Alert charmers as early as possible when they are using `defer()` where it doesn't make sense

The original version of the PR adjusted the various event classes so that the events that shouldn't have `defer()` called didn't have the method at all. This achieves both goals (most IDEs and checkers will detect that a method is being called that isn't defined). However, that was too significant a change, as we could not maintain the pattern for creating custom events and also checks like `isinstance(action_event, ops.EventBase)`.

Instead, we continue with the same practice as we already use for `ActionEvent`, `SecretExpiredEvent`, and `SecretRotateEvent`:
* Set the return type to `typing.NoReturn` - this shows any code after the `defer()` call as unreachable (at least in VS Code) - hopefully there is at least a `return` after the `defer()` call.
* Raises a `RuntimeError` if `defer()` is actually called, which prevents inappropriate `defer()` use in production and will hopefully be caught in tests.

The additional events that get this behaviour are:
* `StopEvent` (I feel that this is the most borderline)
* `RemoveEvent`
* `LifecycleEvent`s

Other options considered (as well as setting the return type):
* Change the signature of `defer()` in the classes that shouldn't use it, e.g. to have a "does_not_support_defer" argument. This will raise a `TypeError` instead of `RuntimeError` or `AttributeError`, with a hint (the name of the argument) as to why it's happened, and this will usually show in IDEs and linters. This provides an earlier warning, which is good, but I feel it's too much a nice hack.
* Call `warnings.warn("x events should not be deferred", RuntimeWarning)` or `warnings.warn("x events should not be deferred", DeprecationWarning)` in the `defer()` in the classes that shouldn't use it. This is safer, but I'm not convinced that it will be noticed. I had thought that unittest/pytest defaulted to something like `-Werror` but that isn't the case. That means that unless someone is running their tests with a similar sort of warnings filter, or is paying close attention to the test output, this is likely to be missed.

I've run the current state of the branch over ~135 Canonical charms' `tox -e unit` and there are no new failures (and the existing failures don't seem to be masking any new issues), so it seems reasonably safe to make this change.