docs: tweak the wording around can_connect() usage (#1123)
`can_connect()` is a point-in-time check, and, if used as a guard for interacting with Pebble, introduces a race condition where Pebble was responding at the `can_connect()` time, but is not when the rest of the work is done.

We already encourage people to handle errors communicating with Pebble when doing the later work (catching `ConnectionError`), which means that most of the time the `can_connect` call is not providing any value (and incurs a small cost). However, the documentation, particularly the example, encourages this pattern, and this has been copied to many other places (charm sources, docs).

This PR adjusts the documentation to be less encouraging of `can_connect()` use as a guard, and changes the example to one where a point-in-time check is wanted (for collecting unit status).