fix(harness): don't error out when attempting attaching storage before begin (#1150)
One minor fix to the `storage` metadata:
[charmcraft.yaml](https://juju.is/docs/sdk/charmcraft-yaml#heading--storage)
and
[metadata.yaml](https://juju.is/docs/sdk/metadata-yaml#heading--storage)
support a `multiple.range` value in the form `<n>+` - this would
previously error out when we attempted to read the metadata. We now
support this (as an alias for `<n>-`, which is an open-ended range
rather than "n or less", and have a test for it.

Adds a minor clarification in the API docs when `Harness.add_storage`
will emit a storage-attached event.

In `begin_with_initial_hooks`, if the storage has already been attached,
then don't try to attach it a second time, but *do* emit the
storage-attached event, since that will not have been done previously.

Make the `_storage_attach` method safe to call multiple times. This
appears to have been the intention based on comments in the code, but,
in practice, attempting to symlink the mount would fail. Now, if the
symlink already exists (and has the same target) then we continue.

Adds a variety of tests for Harness `add_storage` and `attach_storage`
behaviour.

Fixes #1127

---------

Co-authored-by: Ben Hoyt <ben.hoyt@canonical.com>