fix(Harness): remove special-casing for `get_relation` in leader-elected (#1156)
In Juju, using `get_relation()` in a leader-elected event will return
`None` if the relation does not yet exist. This makes Harness consistent
with that behaviour.

This does remove a safety-check for badly written tests - where you mean
for the relation to exist but forget to do an `add_relation()` call
first in the test. However, we don't have any sort of check like this
for any other events, or any other types of relations, and it should be
obvious when the test fails that it's failing to find an expected
relation, and the solution is to set that relation up. I don't think
it's worth having a warning for this specific case (we don't warn in
other similar situations).

This also means we can remove a check in model.py that's a no-op in
production and does work in Harness, which is a bad smell, in my
opinion.

Done:
* [x] [Check with Juju team whether peer-relations are actually
guaranteed to be set up before
leader-elected](https://matrix.to/#/!xzmWHtGpPfVCXKivIh:ubuntu.com/$JUrZzqWEpE5nSQPelnpsG3Mbrd5sGOEiGY989MD-3hs?via=ubuntu.com&via=matrix.org&via=fsfe.org)
* [x] Quick super-tox check to make sure that no-one is somehow relying
on this

Fixes #1153