fix (harness): only inspect the source file if it will be used (#1181)
When experimenting in the REPL it would be convenient to be able to do something like this:

```python
>>> import ops.testing
>>> class MyCharm(ops.CharmBase):
...  pass
... 
>>> h = ops.testing.Harness(MyCharm, meta="{}")
```

This currently fails with `OSError: source code not available`. This PR changes that, so that it will work.

We currently always use `inspect.getfile` to get the source file of the charm, even if metadata is provided to Harness, or if the default metadata would suffice. This PR makes a small adjustment so that we'll only do the `inspect` when needed, and also if it fails we'll fall back to the default metadata.

The change is made for config as well, for consistency.