feat: add pebble.CheckInfo.change_id field (#1197)
This is new with health checks being implemented as changes and tasks
(https://github.com/canonical/pebble/pull/409).

Ops/charms will likely not use this new field. Even once we have the new
`change-updated` event, that has a `.get_change()` method to get the
change directly. But we're adding it to the Pebble client/types for
completeness.