Remove some checks, fix the ruff args, fix a wrongly executable file.
builtin-literals is better handled by a ruff linter (the ones it's finding probably are better converted) and are generally case-by-case.
docstring-first is finding a false positive for a Sphinx odd case - ruff does a bunch of checks, better to leave it there.
debug-statements we have for debug-code so better to disable that entirely for now