chore: add a pre-commit config (#1232)
Adds a basic [pre-commit](https://pre-commit.com/) configuration file.

For anyone that doesn't have pre-commit installed, this will have no
impact. For those that do, running `pre-commit install` in the repo will
install the git pre-commit hooks so that these checks will run on every
commit.

This moves basic checks 'left' so that they are caught as early as
possible.

The config only enables checks from ruff and ones from the 'built-in'
set. We may wish to add others in time, but starting with a simple set
seems best.

Note that pre-commit is already being adopted elsewhere in the charming
community (at least 8 charms, plus scenario and charmcraft), which I'm
broadly in favour of, and I think it would be good for ops to join this
group.