ci: add a workflow that runs the TIOBE quality checks (#1301)
This workflow will run the TIOBE "TICS" quality checks and reports them
through to the TIOBE server.

It's set to run on the first of each month, so that we have regular
updates in TIOBE, but without this hitting every PR (the ruff checks we
have are doing that work).

There's a manual way to do this and there's their packaged action. The
action seems the cleaner choice, so I've gone with that. We still have
to provide the coverage file - we could store that from the normal run
and retrieve it here, which would save the resources of re-running the
tests, but it feels like it is cleaner to keep these completely
separate, which also works well with us splitting that out.