docs: use Sphinx 8 (#1303)
Don't merge!

I tried to used Sphinx 8, but `furo` (style sheet / look) has only
recently begun supporting 8 and there's no release yet. If we wait a
month to a quarter, we should be able move straight to 8.

I was able to get Sphinx 7 to work, would need to clean up the PR if we
decided to merge this.