refactor!: rename Container.exec_mocks to Container.execs and extend mocking (#124)
Firstly, some simple renaming:

* `Container.exec_mocks` becomes `Container.execs`
* `Container.service_status` becomes `Container.service_statuses`
* `ExecOutput` becomes `Exec`

A behaviour change that is a bugfix:

* When a process exits non-zero, the `ExecError` should have the
stdout/stderr if `wait_output` was used (it's not readable afterwards by
the charm, although the Scenario code doesn't enforce that).

Some more substantial changes:

* Provide the ability to get the exact command, the stdin, and all the
other args that the charm used with the process (everything from
os.testing's `ExecArgs`), via a new context attribute `exec_history`,
which is a (default) dict where the key is the container name and the
value is a list of Pebble exec's that have been run.
* Support the same "find the closest match to this prefix" system for
commands as Harness does

We could add more of the functionality that Harness has, but I think
this is a solid subset (I've wanted to be able to get to the stdin in
writing tests myself, and the simple mock matching seems handy). It
should be easy enough to extend in the future without needing API
changes, I think, since this now has both input and output. The key
parts that are missing are properly supporting binary IO and the
execution context.

---------

Co-authored-by: PietroPasotti <pietro.pasotti@canonical.com>