docs: add instructions for moving from 6.x to 7.x (#143)
Since 7.x has so many breaking changes, add an `UPGRADING.md` doc that
lists them all, with before/after examples.

Note that this is *not* the full release notes for 7.x - it doesn't
cover any of the non-breaking changes, other than when they are
incidentally used as part of the examples. We can write release notes
for the release fairly shortly, which can be more comprehensive (but
probably not have as many examples).