docs: expand the reference documentation (#169)
Expands the docstrings so that when used with Sphinx's autodoc they will
fit in with the other ops docs on ops.rtd but also still work
standalone.

The Sphinx autodoc system uses the `__new__` signature in preference to
the `__init__` one, which means that by default all classes that are
using the `_MaxPositionalArgs` class have a `*args, **kwargs` signature,
which is not informative. custom_conf.py monkeypatches Sphinx to work
around this, including tweaking the signature so that the `*` appears in
the correct place for the maximum number of positional arguments.

Also bumps the Sphinx version to align with ops.

---------

Co-authored-by: PietroPasotti <pietro.pasotti@canonical.com>