Fix broken test of registering StatusBase subclass
Previously was catching AttributeError, which must have been raised
historically for missing `name` attribute on StatusBase subclasses.
This, combined with a `type: ignore` directive concealed that the
`register_status` method does not exist -- the attempt to access it
raises an AttributeError. The correct method is `register`, which
raises a TypeError.