Fix broken status_set test
A `type: ignore` directive hid that `status_set` was being called with
the wrong argument types. The intention was to test that an error is
raised on calling with a non-bool argument for `is_app`, however the
status argument should be a string, and was instead a `StatusBase`
class. The test would have still correctly errored out on the `is_app`
argument before this error had an effect at testing time.