Expose settable statuses in type hint of set_status
_ModelBackend.status_set's status argument was previously type hinted
only as str. However, this argument can really only be one of four valid
statuses that juju's set-status will accept. This commit documents this,
and adjusts the definition of StatusBase to also type hint the six
acceptable status names, as well as using these type hints across code
that interacts with StatusBase and status_set.