Revert change to signature of StatusBase.from_name
Since StatusBase.from_name is part of the public API, and is currently
used in several charms, changing the type of the name argument from str
to _StatusName will break users' tests. Furthermore, satisfying the type
checker as to the contents of the strings used may be painful for users,
leading to a proliferation of casts or `type: ignore`s. Therefore, the
type or the name argument has been reverted to str.