ci: update the TIOBE reporting for the changes in coverage calculation (#1367)
We recently changed `tox -e unit` to not produce a coverage report. This
PR adjust the TIOBE workflow to account for that - running `tox -e
coverage` instead.

Since this will now be one of the main uses for that command, I've also
moved the `coverage xml` command into the tox steps rather than being in
the workflow, and adjusted the `.gitignore` accordingly.