feat: expand the secret ID out to the full URI when only given the ID (#1358)
Adjust the rules for getting the 'canonical' form of the Secret URI:
* If the passed ID starts with `secret:`, leave it as-is (unchanged)
* Otherwise, if the model UUID is unknown, add the `secret:` prefix to
end up with the form `secret:id` (unchanged)
* Otherwise, change to the full `secret://{model UUID}/{secret ID` form
(new)

Also adjust the code that calls the canonicalization [sic] method so
that the model UUID is always available in regular ops use. If someone
is creating `SecretInfo` objects themselves then a `None` value for the
UUID is the default, for backwards compatibility (but a warning is
issued in that case).

Fixes #1312.