Fix test failing due to overly specific typing in init
Without type annotations here, the type of {app,unit}_status_to_add is
inferred from the initial assignment of ops.BlockedStatus(...). When
StatusBase.register obscurred the type of all its subclasses, having
them all be typed as StatusBase, this meant that assigning any status to
an untyped variable implicitly typed it as StatusBase. If this is the
intended type of the variable (as it is in this case), it now needs to
be explicitly typed as ops.StatusBase, otherwise the implicit typing
will be the specific status type assigned.