feat: make it an error to call CollectStatusEvent.add_status with error or unknown (#1386)
Add a runtime check to immediately raise an InvalidStatusError if add_status is
called with ErrorStatus or UnknownStatus. Previously these statuses would (in
the case of ErrorStatus) or might (in the case of UnknownStatus) lead to an
InvalidStatusError when it came time to set the status after the collect-status
observer returned. It was never valid to set either status, this just means that
the user will get an immediate exception that points to the line adding the bad
status.