ci: spell-check the code as part of linting (#1388)
Add a `codespell` run as part of the `tox -e lint` environment, to pick
up any spelling errors that might be missed in review. It's also added
to the pre-commit checks, for anyone using those.

The PR also makes 4 corrections, which codespell found. Two are
debatable (whether or not to hyphenate) but two are definite errors.

This adds a new (lint-time) dependency of codespell - it's also used in
the default charmcraft profile, I've used it before, and it has a [good
Snyk score](https://snyk.io/advisor/python/codespell) so it seems ok to
me. I have lightly skimmed the code, but not read it in depth.