ci: run the smoke tests on a schedule (#1387)
We have some basic smoke tests, which are currently the only tests we
regularly run against an actual Juju model. The tests deploy a basic
charm and verify that it goes to `active` status (which happens in an
`install` handler). However, these are currently only run if someone
manually runs `tox -e smoke`. This PR adds running the tests on a
schedule (currently monthly, in the days before we will likely do a
release).

This tests:
* Juju 3.5 (latest stable 3.x - I suggest we *change* this to 3.6 when
that's released, rather than *adding* 3.6). The smoke tests do not
currently work with Juju 2.9 (I'm not sure if it's worth adding that at
this point), but if that changes, it should only require a matrix change
in the workload. The smoke tests use python-libjuju, which doesn't yet
support Juju 4.0, but when that changes, this should also only require a
matrix change.
 * Charmcraft 2.x and 3.x.
 * LXD and MicroK8s.

The smoke test is changed to use a plain `subprocess.run` to pack the
charm, rather than use pytest-operator for this. The `pytest-operator`
approach doesn't work out-of-the-box in the workflow, and it seems
cleaner to just do the pack ourselves.

There's an [example run on my
fork](https://github.com/tonyandrewmeyer/operator/actions/runs/11006199263/job/30560942709).

This could be optimised in the future to store the packed charms and
share them across Juju versions and cloud types, but it doesn't seem
worth doing that for now.

Fixes #1322